+++ src/Aws/CodeDeploy/Resources/codedeploy-2014-10-06.php
@@ -26,6 +26,81 @@ return array (
     'signatureVersion' => 'v4',
     'namespace' => 'CodeDeploy',
     'operations' => array(
+        'AddTagsToOnPremisesInstances' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.AddTagsToOnPremisesInstances',
+                ),
+                'tags' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'Tag',
+                        'type' => 'object',
+                        'properties' => array(
+                            'Key' => array(
+                                'type' => 'string',
+                            ),
+                            'Value' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+                'instanceNames' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'InstanceName',
+                        'type' => 'string',
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'A tag was not specified.',
+                    'class' => 'TagRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified tag was specified in an invalid format.',
+                    'class' => 'InvalidTagException',
+                ),
+                array(
+                    'reason' => 'The maximum allowed number of tags was exceeded.',
+                    'class' => 'TagLimitExceededException',
+                ),
+                array(
+                    'reason' => 'The maximum number of allowed on-premises instances in a single call was exceeded.',
+                    'class' => 'InstanceLimitExceededException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance is not registered.',
+                    'class' => 'InstanceNotRegisteredException',
+                ),
+            ),
+        ),
         'BatchGetApplications' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -68,7 +143,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
             ),
@@ -114,6 +189,47 @@ return array (
                 ),
             ),
         ),
+        'BatchGetOnPremisesInstances' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'BatchGetOnPremisesInstancesOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.BatchGetOnPremisesInstances',
+                ),
+                'instanceNames' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'InstanceName',
+                        'type' => 'string',
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance name was specified in an invalid format.',
+                    'class' => 'InvalidInstanceNameException',
+                ),
+            ),
+        ),
         'CreateApplication' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -153,7 +269,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'An application with the specified name already exists within the AWS user account.',
+                    'reason' => 'An application with the specified name already exists with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationAlreadyExistsException',
                 ),
                 array(
@@ -262,7 +378,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -274,7 +390,7 @@ return array (
                     'class' => 'InvalidDeploymentGroupNameException',
                 ),
                 array(
-                    'reason' => 'The named deployment group does not exist within the AWS user account.',
+                    'reason' => 'The named deployment group does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentGroupDoesNotExistException',
                 ),
                 array(
@@ -290,7 +406,7 @@ return array (
                     'class' => 'InvalidDeploymentConfigNameException',
                 ),
                 array(
-                    'reason' => 'The deployment configuration does not exist within the AWS user account.',
+                    'reason' => 'The deployment configuration does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentConfigDoesNotExistException',
                 ),
                 array(
@@ -350,7 +466,7 @@ return array (
                     'class' => 'DeploymentConfigNameRequiredException',
                 ),
                 array(
-                    'reason' => 'A deployment configuration with the specified name already exists within the AWS user account.',
+                    'reason' => 'A deployment configuration with the specified name already exists with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentConfigAlreadyExistsException',
                 ),
                 array(
@@ -423,6 +539,25 @@ return array (
                         ),
                     ),
                 ),
+                'onPremisesInstanceTagFilters' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'TagFilter',
+                        'type' => 'object',
+                        'properties' => array(
+                            'Key' => array(
+                                'type' => 'string',
+                            ),
+                            'Value' => array(
+                                'type' => 'string',
+                            ),
+                            'Type' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
                 'autoScalingGroups' => array(
                     'type' => 'array',
                     'location' => 'json',
@@ -432,6 +567,7 @@ return array (
                     ),
                 ),
                 'serviceRoleArn' => array(
+                    'required' => true,
                     'type' => 'string',
                     'location' => 'json',
                 ),
@@ -446,7 +582,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -458,12 +594,16 @@ return array (
                     'class' => 'InvalidDeploymentGroupNameException',
                 ),
                 array(
-                    'reason' => 'A deployment group group with the specified name already exists within the AWS user account.',
+                    'reason' => 'A deployment group with the specified name already exists with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentGroupAlreadyExistsException',
                 ),
                 array(
-                    'reason' => 'The Amazon EC2 tag was specified in an invalid format.',
+                    'reason' => 'The tag was specified in an invalid format.',
                     'class' => 'InvalidEC2TagException',
+                ),
+                array(
+                    'reason' => 'The specified tag was specified in an invalid format.',
+                    'class' => 'InvalidTagException',
                 ),
                 array(
                     'reason' => 'The Auto Scaling group was specified in an invalid format or does not exist.',
@@ -474,7 +614,7 @@ return array (
                     'class' => 'InvalidDeploymentConfigNameException',
                 ),
                 array(
-                    'reason' => 'The deployment configuration does not exist within the AWS user account.',
+                    'reason' => 'The deployment configuration does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentConfigDoesNotExistException',
                 ),
                 array(
@@ -638,6 +778,44 @@ return array (
                 ),
             ),
         ),
+        'DeregisterOnPremisesInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.DeregisterOnPremisesInstance',
+                ),
+                'instanceName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance name was specified in an invalid format.',
+                    'class' => 'InvalidInstanceNameException',
+                ),
+            ),
+        ),
         'GetApplication' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -677,7 +855,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
             ),
@@ -754,7 +932,7 @@ return array (
             ),
             'errorResponses' => array(
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -766,7 +944,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The named revision does not exist within the AWS user account.',
+                    'reason' => 'The named revision does not exist with the applicable IAM user or AWS account.',
                     'class' => 'RevisionDoesNotExistException',
                 ),
                 array(
@@ -816,7 +994,7 @@ return array (
                     'class' => 'InvalidDeploymentIdException',
                 ),
                 array(
-                    'reason' => 'The deployment does not exist within the AWS user account.',
+                    'reason' => 'The deployment does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentDoesNotExistException',
                 ),
             ),
@@ -860,7 +1038,7 @@ return array (
                     'class' => 'DeploymentConfigNameRequiredException',
                 ),
                 array(
-                    'reason' => 'The deployment configuration does not exist within the AWS user account.',
+                    'reason' => 'The deployment configuration does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentConfigDoesNotExistException',
                 ),
             ),
@@ -911,7 +1089,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -923,7 +1101,7 @@ return array (
                     'class' => 'InvalidDeploymentGroupNameException',
                 ),
                 array(
-                    'reason' => 'The named deployment group does not exist within the AWS user account.',
+                    'reason' => 'The named deployment group does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentGroupDoesNotExistException',
                 ),
             ),
@@ -966,7 +1144,7 @@ return array (
                     'class' => 'DeploymentIdRequiredException',
                 ),
                 array(
-                    'reason' => 'The deployment does not exist within the AWS user account.',
+                    'reason' => 'The deployment does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentDoesNotExistException',
                 ),
                 array(
@@ -980,6 +1158,48 @@ return array (
                 array(
                     'reason' => 'The specified instance does not exist in the deployment group.',
                     'class' => 'InstanceDoesNotExistException',
+                ),
+            ),
+        ),
+        'GetOnPremisesInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'GetOnPremisesInstanceOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.GetOnPremisesInstance',
+                ),
+                'instanceName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance is not registered.',
+                    'class' => 'InstanceNotRegisteredException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance name was specified in an invalid format.',
+                    'class' => 'InvalidInstanceNameException',
                 ),
             ),
         ),
@@ -1038,7 +1258,7 @@ return array (
             ),
             'errorResponses' => array(
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -1188,7 +1408,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -1242,7 +1462,7 @@ return array (
                     'class' => 'DeploymentIdRequiredException',
                 ),
                 array(
-                    'reason' => 'The deployment does not exist within the AWS user account.',
+                    'reason' => 'The deployment does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentDoesNotExistException',
                 ),
                 array(
@@ -1341,7 +1561,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -1349,7 +1569,7 @@ return array (
                     'class' => 'InvalidDeploymentGroupNameException',
                 ),
                 array(
-                    'reason' => 'The named deployment group does not exist within the AWS user account.',
+                    'reason' => 'The named deployment group does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentGroupDoesNotExistException',
                 ),
                 array(
@@ -1363,6 +1583,70 @@ return array (
                 array(
                     'reason' => 'The specified deployment status doesn\'t exist or cannot be determined.',
                     'class' => 'InvalidDeploymentStatusException',
+                ),
+                array(
+                    'reason' => 'The next token was specified in an invalid format.',
+                    'class' => 'InvalidNextTokenException',
+                ),
+            ),
+        ),
+        'ListOnPremisesInstances' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'ListOnPremisesInstancesOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.ListOnPremisesInstances',
+                ),
+                'registrationStatus' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'tagFilters' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'TagFilter',
+                        'type' => 'object',
+                        'properties' => array(
+                            'Key' => array(
+                                'type' => 'string',
+                            ),
+                            'Value' => array(
+                                'type' => 'string',
+                            ),
+                            'Type' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The registration status was specified in an invalid format.',
+                    'class' => 'InvalidRegistrationStatusException',
+                ),
+                array(
+                    'reason' => 'The specified tag filter was specified in an invalid format.',
+                    'class' => 'InvalidTagFilterException',
                 ),
                 array(
                     'reason' => 'The next token was specified in an invalid format.',
@@ -1446,7 +1730,7 @@ return array (
             ),
             'errorResponses' => array(
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -1468,6 +1752,140 @@ return array (
                 array(
                     'reason' => 'The revision was specified in an invalid format.',
                     'class' => 'InvalidRevisionException',
+                ),
+            ),
+        ),
+        'RegisterOnPremisesInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.RegisterOnPremisesInstance',
+                ),
+                'instanceName' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'iamUserArn' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The specified on-premises instance name is already registered.',
+                    'class' => 'InstanceNameAlreadyRegisteredException',
+                ),
+                array(
+                    'reason' => 'The specified IAM user ARN is already registered with an on-premises instance.',
+                    'class' => 'IamUserArnAlreadyRegisteredException',
+                ),
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'An IAM user ARN was not specified.',
+                    'class' => 'IamUserArnRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance name was specified in an invalid format.',
+                    'class' => 'InvalidInstanceNameException',
+                ),
+                array(
+                    'reason' => 'The IAM user ARN was specified in an invalid format.',
+                    'class' => 'InvalidIamUserArnException',
+                ),
+            ),
+        ),
+        'RemoveTagsFromOnPremisesInstances' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'CodeDeploy_20141006.RemoveTagsFromOnPremisesInstances',
+                ),
+                'tags' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'Tag',
+                        'type' => 'object',
+                        'properties' => array(
+                            'Key' => array(
+                                'type' => 'string',
+                            ),
+                            'Value' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
+                'instanceNames' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'InstanceName',
+                        'type' => 'string',
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'An on-premises instance name was not specified.',
+                    'class' => 'InstanceNameRequiredException',
+                ),
+                array(
+                    'reason' => 'A tag was not specified.',
+                    'class' => 'TagRequiredException',
+                ),
+                array(
+                    'reason' => 'The specified tag was specified in an invalid format.',
+                    'class' => 'InvalidTagException',
+                ),
+                array(
+                    'reason' => 'The maximum allowed number of tags was exceeded.',
+                    'class' => 'TagLimitExceededException',
+                ),
+                array(
+                    'reason' => 'The maximum number of allowed on-premises instances in a single call was exceeded.',
+                    'class' => 'InstanceLimitExceededException',
+                ),
+                array(
+                    'reason' => 'The specified on-premises instance is not registered.',
+                    'class' => 'InstanceNotRegisteredException',
                 ),
             ),
         ),
@@ -1504,7 +1922,7 @@ return array (
                     'class' => 'DeploymentIdRequiredException',
                 ),
                 array(
-                    'reason' => 'The deployment does not exist within the AWS user account.',
+                    'reason' => 'The deployment does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentDoesNotExistException',
                 ),
                 array(
@@ -1561,11 +1979,11 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'An application with the specified name already exists within the AWS user account.',
+                    'reason' => 'An application with the specified name already exists with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationAlreadyExistsException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
             ),
@@ -1636,6 +2054,25 @@ return array (
                         ),
                     ),
                 ),
+                'onPremisesInstanceTagFilters' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'TagFilter',
+                        'type' => 'object',
+                        'properties' => array(
+                            'Key' => array(
+                                'type' => 'string',
+                            ),
+                            'Value' => array(
+                                'type' => 'string',
+                            ),
+                            'Type' => array(
+                                'type' => 'string',
+                            ),
+                        ),
+                    ),
+                ),
                 'autoScalingGroups' => array(
                     'type' => 'array',
                     'location' => 'json',
@@ -1659,7 +2096,7 @@ return array (
                     'class' => 'InvalidApplicationNameException',
                 ),
                 array(
-                    'reason' => 'The application does not exist within the AWS user account.',
+                    'reason' => 'The application does not exist with the applicable IAM user or AWS account.',
                     'class' => 'ApplicationDoesNotExistException',
                 ),
                 array(
@@ -1667,7 +2104,7 @@ return array (
                     'class' => 'InvalidDeploymentGroupNameException',
                 ),
                 array(
-                    'reason' => 'A deployment group group with the specified name already exists within the AWS user account.',
+                    'reason' => 'A deployment group with the specified name already exists with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentGroupAlreadyExistsException',
                 ),
                 array(
@@ -1675,8 +2112,16 @@ return array (
                     'class' => 'DeploymentGroupNameRequiredException',
                 ),
                 array(
-                    'reason' => 'The Amazon EC2 tag was specified in an invalid format.',
+                    'reason' => 'The named deployment group does not exist with the applicable IAM user or AWS account.',
+                    'class' => 'DeploymentGroupDoesNotExistException',
+                ),
+                array(
+                    'reason' => 'The tag was specified in an invalid format.',
                     'class' => 'InvalidEC2TagException',
+                ),
+                array(
+                    'reason' => 'The specified tag was specified in an invalid format.',
+                    'class' => 'InvalidTagException',
                 ),
                 array(
                     'reason' => 'The Auto Scaling group was specified in an invalid format or does not exist.',
@@ -1687,7 +2132,7 @@ return array (
                     'class' => 'InvalidDeploymentConfigNameException',
                 ),
                 array(
-                    'reason' => 'The deployment configuration does not exist within the AWS user account.',
+                    'reason' => 'The deployment configuration does not exist with the applicable IAM user or AWS account.',
                     'class' => 'DeploymentConfigDoesNotExistException',
                 ),
                 array(
@@ -1698,6 +2143,10 @@ return array (
         ),
     ),
     'models' => array(
+        'EmptyOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+        ),
         'BatchGetApplicationsOutput' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1845,6 +2294,52 @@ return array (
                 ),
             ),
         ),
+        'BatchGetOnPremisesInstancesOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'instanceInfos' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'InstanceInfo',
+                        'type' => 'object',
+                        'properties' => array(
+                            'instanceName' => array(
+                                'type' => 'string',
+                            ),
+                            'iamUserArn' => array(
+                                'type' => 'string',
+                            ),
+                            'instanceArn' => array(
+                                'type' => 'string',
+                            ),
+                            'registerTime' => array(
+                                'type' => 'string',
+                            ),
+                            'deregisterTime' => array(
+                                'type' => 'string',
+                            ),
+                            'tags' => array(
+                                'type' => 'array',
+                                'items' => array(
+                                    'name' => 'Tag',
+                                    'type' => 'object',
+                                    'properties' => array(
+                                        'Key' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'Value' => array(
+                                            'type' => 'string',
+                                        ),
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
         'CreateApplicationOutput' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1884,10 +2379,6 @@ return array (
                     'location' => 'json',
                 ),
             ),
-        ),
-        'EmptyOutput' => array(
-            'type' => 'object',
-            'additionalProperties' => true,
         ),
         'DeleteDeploymentGroupOutput' => array(
             'type' => 'object',
@@ -2191,6 +2682,24 @@ return array (
                                 ),
                             ),
                         ),
+                        'onPremisesInstanceTagFilters' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'TagFilter',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Key' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'Value' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'Type' => array(
+                                        'type' => 'string',
+                                    ),
+                                ),
+                            ),
+                        ),
                         'autoScalingGroups' => array(
                             'type' => 'array',
                             'items' => array(
@@ -2305,6 +2814,48 @@ return array (
                                         'type' => 'string',
                                     ),
                                     'status' => array(
+                                        'type' => 'string',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
+        'GetOnPremisesInstanceOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'instanceInfo' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'instanceName' => array(
+                            'type' => 'string',
+                        ),
+                        'iamUserArn' => array(
+                            'type' => 'string',
+                        ),
+                        'instanceArn' => array(
+                            'type' => 'string',
+                        ),
+                        'registerTime' => array(
+                            'type' => 'string',
+                        ),
+                        'deregisterTime' => array(
+                            'type' => 'string',
+                        ),
+                        'tags' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'Tag',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Key' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'Value' => array(
                                         'type' => 'string',
                                     ),
                                 ),
@@ -2453,6 +3004,24 @@ return array (
                     'location' => 'json',
                     'items' => array(
                         'name' => 'DeploymentId',
+                        'type' => 'string',
+                    ),
+                ),
+                'nextToken' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+        ),
+        'ListOnPremisesInstancesOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'instanceNames' => array(
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'InstanceName',
                         'type' => 'string',
                     ),
                 ),